Specifying iterations and convergence tolerance. Moving tests to their own module.
